+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@
 #spring.datasource.driverClassName=com.microsoft.sqlserver.jdbc.SQLServerDriver
 spring.datasource.url=jdbc:mysql://localhost:3306/db_datn
 spring.datasource.username=root
-spring.datasource.password=12345678
+spring.datasource.password=khoibi2003
 
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.show-sql=true
